+++ fastkml/base.py
@@ -16,8 +16,6 @@
 
 """Abstract base classes"""
 import logging
-from typing import Any
-from typing import Callable
 from typing import Optional
 from typing import Tuple
 from typing import cast
@@ -26,130 +24,10 @@ from fastkml import config
 from fastkml.types import Element
 from fastkml.types import KmlObjectMap
 
+from .helpers import o_from_attr
+from .helpers import o_to_attr
+
 logger = logging.getLogger(__name__)
-
-
-def o_to_attr(
-    obj: object,
-    element: Element,
-    kml_attr: str,
-    obj_attr: str,
-    required: bool,
-    **kwargs: Any,
-) -> None:
-    """Set an attribute on an KML Element from an object attribute."""
-    attribute = getattr(obj, obj_attr)
-    if attribute:
-        element.set(kml_attr, str(attribute))
-    elif required:
-        logger.warning(
-            "Required attribute '%s' for '%s' missing.",
-            obj_attr,
-            obj.__class__.__name__,
-        )
-
-
-def o_from_attr(
-    obj: object,
-    element: Element,
-    kml_attr: str,
-    obj_attr: str,
-    required: bool,
-    **kwargs: Any,
-) -> None:
-    """Set an attribute on self from an KML attribute."""
-    attribute = element.get(kml_attr)
-    if attribute:
-        setattr(obj, obj_attr, attribute)
-    elif required:
-        logger.warning(
-            "Required attribute '%s' for '%s' missing.",
-            kml_attr,
-            obj.__class__.__name__,
-        )
-
-
-def o_int_from_attr(
-    obj: object,
-    element: Element,
-    kml_attr: str,
-    obj_attr: str,
-    required: bool,
-    **kwargs: Any,
-) -> None:
-    """Set an attribute on self from an KML attribute."""
-    try:
-        attribute = int(element.get(kml_attr))
-    except (ValueError, TypeError):
-        attribute = None
-    if attribute is not None:
-        setattr(obj, obj_attr, attribute)
-    elif required:
-        logger.warning(
-            "Required attribute '%s' for '%s' missing.",
-            kml_attr,
-            obj.__class__.__name__,
-        )
-
-
-def o_from_subelement_text(
-    obj: object,
-    element: Element,
-    kml_attr: str,
-    obj_attr: str,
-    required: bool,
-    validator: Optional[Callable[..., bool]] = None,
-    **kwargs: Any,
-) -> None:
-    """Set an attribute on self from the text of a SubElement."""
-    elem = element.find(f"{obj.ns}{kml_attr}")  # type: ignore[attr-defined]
-    if elem is not None:
-        if validator is not None and not validator(elem.text):
-            logger.warning(
-                "Invalid value for attribute '%s' for '%s'",
-                kml_attr,
-                obj.__class__.__name__,
-            )
-        else:
-            setattr(obj, obj_attr, elem.text)
-    elif required:  # type: ignore[unreachable]
-        logger.warning(
-            "Required attribute '%s' for '%s' missing.",
-            kml_attr,
-            obj.__class__.__name__,
-        )
-
-
-def o_to_subelement_text(
-    obj: object,
-    element: Element,
-    kml_attr: str,
-    obj_attr: str,
-    required: bool,
-    validator: Optional[Callable[..., bool]] = None,
-    **kwargs: Any,
-) -> None:
-    """Set the text of a SubElement from an object attribute."""
-    attribute = getattr(obj, obj_attr)
-    if attribute:
-        if validator is not None and not validator(attribute):
-            logger.warning(
-                "Invalid value for attribute '%s' for '%s'",
-                obj_attr,
-                obj.__class__.__name__,
-            )
-        else:
-            elem = config.etree.SubElement(
-                element,  # type: ignore[arg-type]
-                f"{obj.ns}{kml_attr}",  # type: ignore[attr-defined]
-            )
-            elem.text = str(attribute)
-    elif required:
-        logger.warning(
-            "Required attribute '%s' for '%s' missing.",
-            obj_attr,
-            obj.__class__.__name__,
-        )
 
 
 class _XMLObject:
