+++ setup.py
@@ -7,12 +7,12 @@ from setuptools.command.test import test as TestCommand
 
 
 class PyTest(TestCommand):
-    def finalize_options(self):
+    def finalize_options(self) -> None:
         TestCommand.finalize_options(self)
         self.test_args = []
         self.test_suite = True
 
-    def run_tests(self):
+    def run_tests(self) -> None:
         # import here, cause outside the eggs aren't loaded
         import pytest
